refactor: Simplify header component structure by delegating to specialized components
- Refactor HeaderComponents.kt to act as a facade for header and time display components, improving code organization and maintainability.
- Remove legacy CalendarHeader and DisplayTimeInIran composables, directing usage to the new HeaderSection for better clarity and consistency.
